Rename config.json to config.js
Spent too long trying to figure out why next-stream-page can't handle
json files anymore. This fixes it.
