+++ tour-tip/config.js
